update to read web layer from portal
Manual projects list and opportunity projects list are now hosted on MTC web portal, which allows users to edit online.
The script is updated to read the feature layer from the portal directly, which require VPN connection and connection to MTC web portal.
If users have no access to these connections, please use the file in the previous commit.
Only revised is the machine. It is now set up to run on satmodel, instead of AWS.
A few minor error corrections - no impact on the results